+++ tests/src/TestCase.php
@@ -2,11 +2,21 @@
 
 namespace TomatoPHP\LaravelLogstash\Tests;
 
+use Illuminate\Support\Facades\Http;
 use Orchestra\Testbench\TestCase as BaseTestCase;
 use TomatoPHP\LaravelLogstash\LaravelLogstashServiceProvider;
 
 abstract class TestCase extends BaseTestCase
 {
+    public const LOGSTASH_URL = 'https://logstash.test:8080';
+
+    protected function setUp(): void
+    {
+        parent::setUp();
+
+        Http::preventStrayRequests();
+    }
+
     protected function getPackageProviders($app): array
     {
         return [
@@ -14,9 +24,10 @@ abstract class TestCase extends BaseTestCase
         ];
     }
 
-    public function getEnvironmentSetUp($app): void
+    protected function defineEnvironment($app): void
     {
-        $app['config']->set('logging.channels', 'logstash');
+        $app['config']->set('app.name', 'Testing');
         $app['config']->set('queue.default', 'sync');
+        $app['config']->set('laravel-logstash.url', self::LOGSTASH_URL);
     }
 }